Fix up tests, remove unused code.
